requirements.txt
Flask==2.3.2
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.0
tensorflow==2.16.1
PyYAML==6.0.1
python-dotenv==1.0.1
flask-cors==4.0.0 # Added for CORS support